+++ R/crew_client.R
@@ -22,6 +22,12 @@
 #' @param seconds_timeout Number of seconds until timing
 #'   out while waiting for certain synchronous operations to complete,
 #'   such as checking `mirai::status()`.
+#' @param serialization Either `NULL` (default) or an object produced by
+#'   [mirai::serial_config()] to control the serialization
+#'   of data sent to workers. This can help with either more efficient
+#'   data transfers or to preserve attributes of otherwise
+#'   non-exportable objects (such as `torch` tensors or `arrow` tables).
+#'   See `?mirai::serial_config` for details.
 #' @param retry_tasks Deprecated on 2025-01-13 (`crew` version 0.10.2.9002).
 #' @examples
 #' if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
@@ -38,6 +44,7 @@ crew_client <- function(
   tls = crew::crew_tls(),
   tls_enable = NULL,
   tls_config = NULL,
+  serialization = NULL,
   seconds_interval = 1,
   seconds_timeout = 60,
   retry_tasks = NULL
@@ -90,6 +97,7 @@ crew_client <- function(
     host = host,
     port = port,
     tls = tls,
+    serialization = serialization,
     seconds_interval = seconds_interval,
     seconds_timeout = seconds_timeout,
     relay = crew_relay()
@@ -117,12 +125,14 @@ crew_class_client <- R6::R6Class(
     .host = NULL,
     .port = NULL,
     .tls = NULL,
+    .serialization = NULL,
     .seconds_interval = NULL,
     .seconds_timeout = NULL,
     .relay = NULL,
     .started = FALSE,
     .url = NULL,
     .profile = NULL,
+    .condition = NULL,
     .client = NULL, # TODO: remove if/when the dispatcher becomes a thread.
     .dispatcher = NULL # TODO: remove if/when the dispatcher becomes a thread.
   ),
@@ -138,6 +148,10 @@ crew_class_client <- R6::R6Class(
     #' @field tls See [crew_client()].
     tls = function() {
       .subset2(private, ".tls")
+    },
+    #' @field serialization See [crew_client()].
+    serialization = function() {
+      .subset2(private, ".serialization")
     },
     #' @field seconds_interval See [crew_client()].
     seconds_interval = function() {
@@ -164,6 +178,10 @@ crew_class_client <- R6::R6Class(
     profile = function() {
       .subset2(private, ".profile")
     },
+    #' @field condition Condition variable of the client.
+    condition = function() {
+      .subset2(private, ".condition")
+    },
     #' @field client Process ID of the local process running the client.
     client = function() {
       .subset2(private, ".client")
@@ -179,6 +197,7 @@ crew_class_client <- R6::R6Class(
     #' @param host Argument passed from [crew_client()].
     #' @param port Argument passed from [crew_client()].
     #' @param tls Argument passed from [crew_client()].
+    #' @param serialization Argument passed from [crew_client()].
     #' @param seconds_interval Argument passed from [crew_client()].
     #' @param seconds_timeout Argument passed from [crew_client()].
     #' @param relay Argument passed from [crew_client()].
@@ -193,6 +212,7 @@ crew_class_client <- R6::R6Class(
       host = NULL,
       port = NULL,
       tls = NULL,
+      serialization = NULL,
       seconds_interval = NULL,
       seconds_timeout = NULL,
       relay = NULL
@@ -200,8 +220,12 @@ crew_class_client <- R6::R6Class(
       private$.host <- host
       private$.port <- port
       private$.tls <- tls
+      private$.serialization <- serialization
       private$.seconds_interval <- seconds_interval
       private$.seconds_timeout <- seconds_timeout
+      # Creating the CV here instead of as a default R6 field value
+      # somehow appeases covr and R CMD check.
+      private$.condition <- nanonext::cv()
       private$.relay <- relay
     },
     #' @description Validate the client.
@@ -245,8 +269,21 @@ crew_class_client <- R6::R6Class(
       )
       crew_assert(private$.seconds_timeout >= private$.seconds_interval)
       crew_assert(inherits(private$.relay, "crew_class_relay"))
+      if_any(
+        is.null(private$.serialization),
+        NULL,
+        crew_assert(is.list(private$.serialization))
+      )
       private$.relay$validate()
       invisible()
+    },
+    #' @description Register the client as started.
+    #' @details Exported to implement the sequential controller.
+    #'   Only meant to be called manually inside the client or
+    #'   the sequential controller.
+    #' @return `NULL` (invisibly).
+    set_started = function() {
+      private$.started <- TRUE
     },
     #' @description Start listening for workers on the available sockets.
     #' @return `NULL` (invisibly).
@@ -259,6 +296,7 @@ crew_class_client <- R6::R6Class(
         url = private$.tls$url(host = private$.host, port = private$.port),
         dispatcher = TRUE,
         seed = NULL,
+        serial = private$.serialization,
         tls = private$.tls$client(),
         pass = private$.tls$password,
         .compute = private$.profile
@@ -273,7 +311,11 @@ crew_class_client <- R6::R6Class(
         private$.dispatcher <- ps::ps_handle(pid = pid)
       }
       # End dispatcher code.
-      private$.relay$set_from(self$condition())
+      private$.condition <- mirai::nextget(
+        x = "cv",
+        .compute = private$.profile
+      )
+      private$.relay$set_from(.subset2(private, ".condition"))
       private$.relay$start()
       private$.started <- TRUE
       invisible()
@@ -285,8 +327,11 @@ crew_class_client <- R6::R6Class(
       if (!isTRUE(private$.started)) {
         return(invisible())
       }
-      mirai::daemons(n = 0L, .compute = private$.profile)
+      if (!is.null(private$.profile)) {
+        mirai::daemons(n = 0L, .compute = private$.profile)
+      }
       private$.profile <- NULL
+      private$.condition <- nanonext::cv()
       private$.relay$terminate()
       private$.url <- NULL
       private$.started <- FALSE
@@ -323,28 +368,10 @@ crew_class_client <- R6::R6Class(
       # End dispatcher checks.
       invisible()
     },
-    #' @description Get the `nanonext` condition variable which tasks signal
-    #'   on resolution.
-    #' @return The `nanonext` condition variable which tasks signal
-    #'   on resolution. The return value is `NULL` if the client
-    #'   is not running.
-    condition = function() {
-      profile <- .subset2(private, ".profile")
-      if (is.null(profile)) {
-        NULL
-      } else {
-        mirai::nextget(x = "cv", .compute = profile)
-      }
-    },
     #' @description Get the true value of the `nanonext` condition variable.
     #' @return The value of the `nanonext` condition variable.
     resolved = function() {
-      condition <- .subset2(self, "condition")()
-      if (is.null(condition)) {
-        0L
-      } else {
-        nanonext::cv_value(condition)
-      }
+      nanonext::cv_value(.subset2(private, ".condition"))
     },
     #' @description Internal function:
     #'   return the `mirai` status of the compute profile.