+++ R/utils_packages.R
@@ -1,21 +1,21 @@
 load_packages <- function(packages, library) {
-  out <- suppressPackageStartupMessages(
-    lapply(
-      packages,
-      require,
-      lib.loc = library,
-      quietly = TRUE,
-      character.only = TRUE
-    )
+  out <- lapply(
+    packages,
+    require,
+    lib.loc = library,
+    quietly = TRUE,
+    character.only = TRUE
   )
   out <- as.logical(unlist(out))
-  msg <- paste(
-    "could not find packages",
-    paste(packages[!out], collapse = ", "),
-    "in library paths",
-    paste(library, collapse = ", ")
-  )
-  crew_assert(all(out), message = msg)
+  if (!all(out)) {
+    msg <- paste(
+      "could not find packages",
+      paste(packages[!out], collapse = ", "),
+      "in library paths",
+      paste(library, collapse = ", ")
+    )
+    crew_error(message = msg)
+  }
 }
 
 package_installed <- function(package) {